azure-functions/requirements.txt
apify-client==1.8.1
google-genai==1.13.0
firebase-admin==6.5.0
azure-functions==1.21.3
